Perfil ya se crea, falta asociaciones!

+++ src/main/java/com/bolsaTrabajo/repositories/CompanyRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository("companyRepository")
-public interface CompanyRepository extends JpaRepository<Company,Long> {
+public interface CompanyRepository extends JpaRepository<Company,Integer> {
     Company findByUsername(String username);
-    Company findById(long id);
+    Company findById(Long id);
 }
